api/requirements.txt
absl-py==1.4.0
anyio==3.7.0
boto3==1.26.149
botocore==1.29.149
cachetools==5.3.1
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.10.0
exceptiongroup==1.1.1
fastapi==0.95.2
filelock==3.12.0
fire==0.5.0
fonttools==4.39.4
fsspec==2023.5.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.19.1
google-auth-oauthlib==1.0.0
grpcio==1.54.2
h11==0.14.0
httptools==0.5.0
huggingface-hub==0.15.1
idna==2.10
importlib-metadata==6.6.0
importlib-resources==5.12.0
Jinja2==3.1.2
jmespath==1.0.1
kiwisolver==1.4.4
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
protobuf==4.23.2
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybboxes==0.1.6
pydantic==1.10.8
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
roboflow==1.0.9
rsa==4.9
s3transfer==0.6.1
sahi==0.11.14
scipy==1.10.1
seaborn==0.12.2
sentry-sdk==1.24.0
shapely==2.0.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.0
termcolor==2.3.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
typing_extensions==4.6.2
tzdata==2023.3
ultralytics==8.0.109
urllib3==1.26.16
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.3
Werkzeug==2.3.5
wget==3.2
yolov5==7.0.12
zipp==3.15.0
